Updates with suggestions
- Introduced Vtol.Debugging namespace
- Added Vtol.Debugging folder
- Created VtolLogger class
- Updated `Lazy<List<PriorityConnectionFilter>>` initialization
- Changed filter priority from `int` to `double`
- Added message to log-file when user tries to add a connection and it already exists

+++ Voxel Tycoon Open Library/StorageNetwork/PriorityConnectionFilter.cs
@@ -7,7 +7,7 @@ namespace VTOL.StorageNetwork
 	/// </summary>
 	internal struct PriorityConnectionFilter : IComparable<PriorityConnectionFilter>
 	{
-		public PriorityConnectionFilter(IConnectionFilter connectionFilter, int priority)
+		public PriorityConnectionFilter(IConnectionFilter connectionFilter, double priority)
 		{
 			ConnectionFilter = connectionFilter;
 			Priority = priority;
@@ -17,7 +17,7 @@ namespace VTOL.StorageNetwork
 		/// The priority of the filter.
 		/// </summary>
 		/// <remarks>Filters with a higher priority will be executed after filters with a lower priority. Meaning the alterations made by a filter with a higher priority cannot be overwritten by a filter with a lower priority.</remarks>
-		public int Priority { get; }
+		public double Priority { get; }
 
 		/// <summary>
 		/// The class with the filter logic.
